requirements.txt
beautifulsoup4
requests
django
djangorestframework
django-filter
